requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
colorama==0.4.6
cryptography==44.0.2
Cython==3.0.12
decorator==5.2.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.18.0
fire==0.7.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
protobuf==3.20.2
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pypdfium2==4.30.1
PySocks==1.7.1
python-bidi==0.6.6
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
selenium==4.30.0
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stringzilla==3.12.3
sympy==1.13.1
termcolor==2.5.0
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
